Add instruction details to cstest logging (#2787)
* Update cstest.py logging

* Update cstest.c logging

* Change exception to log+return in TestExpected.compare

The prior change in TestCase.test converts the exception to
TestResult.FAILED which doesn't preserve the status here.

* Restore calling sequence in TestCase.test

The transformation didn't preserve error reporting. It also is
unnecessary now that TestExpected.compare returns a TestResult again.

* Convert exception to return for length check in TestExpected.compare

The callsite in TestCase.test no longer converts exceptions to a
TestResult.

* Rename local variable 'postfix' to 'prefix' in test_input_stringify

This better reflects what it actually is.