Update read/written registers for x86 string instructions (#2790)
* stosb/w/d/q

stosq only reads/writes rcx when there is a rep prefix.

With regard to reading 'es' in 64-bit mode, the SDM says for stosd

  For legacy mode, store EAX at address ES:(E)DI; For 64-bit mode store
  EAX at address RDI or EDI.

* lodsb/w/d/q

With regard to reading 'ds' in 64-bit mode, the SDM says for lodsd

  For legacy mode, Load dword at address DS:(E)SI into EAX. For 64-bit
  mode load dword at address (R)SI into EAX.

* scasb/w/d/q

X86_REG_ES was removed from the explicit set of read registers because
it was causing duplicate entries in 'detail->regs_read'.

With regard to reading 'es' in 64-bit mode, the SDM says for scasd

  Compare EAX with doubleword at ES:(E)DI or RDI then set status flags.

* cmpsb/w/d/q

The 'es' register is implicitly handled correctly. The 'ds' register
should only be read from in 16- and 32-bit mode. From the SDM for cmpsd:

  For legacy mode, compare dword at address DS:(E)SI with dword at
  address ES:(E)DI; For 64-bit mode compare dword at address (R|E)SI
  with dword at address (R|E)DI. The status flags are set accordingly.

* movsb/w/d/q

The 'es' register is implicitly handled correctly. The 'ds' register
should only be read from in 16- and 32-bit mode. From the SDM for movsd:

  For legacy mode, move dword from address DS:(E)SI to ES:(E)DI. For
  64-bit mode move dword from address (R|E)SI to (R|E)DI.

* insb/w/d/q

X86_REG_EDI was removed from the explicit set of read registers because
it was causing duplicate entries in 'detail->regs_read'.

The 'ds' register should only be read from in 16- and 32-bit mode. From
the SDM for insd:

  Input doubleword from I/O port specified in DX into memory location
  specified in ES:(E)DI or RDI.

* outsb/w/d/q

The 'ds' register should only be read from in 16- and 32-bit mode. From
the SDM for outsd:

  Output word from memory location specified in DS:(E)SI or RSI to I/O
  port specified in DX.